extract out progress reporting
