+++ ghcide/src/Development/IDE/Core/Shake.hs
@@ -85,35 +85,35 @@ import           Control.DeepSeq
 import           Control.Monad.Extra
 import           Control.Monad.IO.Class
 import           Control.Monad.Reader
-import qualified Control.Monad.STM                    as STM
 import           Control.Monad.Trans.Maybe
-import qualified Data.ByteString.Char8                as BS
+import qualified Data.ByteString.Char8                  as BS
 import           Data.Dynamic
-import qualified Data.HashMap.Strict                  as HMap
+import qualified Data.HashMap.Strict                    as HMap
 import           Data.Hashable
-import           Data.List.Extra                      (partition, takeEnd)
-import           Data.Map.Strict                      (Map)
-import qualified Data.Map.Strict                      as Map
+import           Data.List.Extra                        (partition, takeEnd)
+import           Data.Map.Strict                        (Map)
+import qualified Data.Map.Strict                        as Map
 import           Data.Maybe
-import qualified Data.Set                             as Set
-import qualified Data.SortedList                      as SL
-import qualified Data.Text                            as T
+import qualified Data.Set                               as Set
+import qualified Data.SortedList                        as SL
+import qualified Data.Text                              as T
 import           Data.Time
 import           Data.Traversable
 import           Data.Tuple.Extra
 import           Data.Typeable
 import           Data.Unique
-import           Data.Vector                          (Vector)
-import qualified Data.Vector                          as Vector
+import           Data.Vector                            (Vector)
+import qualified Data.Vector                            as Vector
 import           Development.IDE.Core.Debouncer
 import           Development.IDE.Core.PositionMapping
+import           Development.IDE.Core.ProgressReporting
 import           Development.IDE.Core.RuleTypes
 import           Development.IDE.Core.Tracing
-import           Development.IDE.GHC.Compat           (NameCacheUpdater (..),
-                                                       upNameCache)
-import           Development.IDE.GHC.Orphans          ()
-import           Development.IDE.Graph                hiding (ShakeValue)
-import qualified Development.IDE.Graph                as Shake
+import           Development.IDE.GHC.Compat             (NameCacheUpdater (..),
+                                                         upNameCache)
+import           Development.IDE.GHC.Orphans            ()
+import           Development.IDE.Graph                  hiding (ShakeValue)
+import qualified Development.IDE.Graph                  as Shake
 import           Development.IDE.Graph.Classes
 import           Development.IDE.Graph.Database
 import           Development.IDE.Graph.Rule
@@ -122,17 +122,17 @@ import           Development.IDE.Types.Diagnostics
 import           Development.IDE.Types.Exports
 import           Development.IDE.Types.KnownTargets
 import           Development.IDE.Types.Location
-import           Development.IDE.Types.Logger         hiding (Priority)
-import qualified Development.IDE.Types.Logger         as Logger
+import           Development.IDE.Types.Logger           hiding (Priority)
+import qualified Development.IDE.Types.Logger           as Logger
 import           Development.IDE.Types.Options
 import           Development.IDE.Types.Shake
 import           GHC.Generics
 import           Language.LSP.Diagnostics
-import qualified Language.LSP.Server                  as LSP
+import qualified Language.LSP.Server                    as LSP
 import           Language.LSP.Types
-import qualified Language.LSP.Types                   as LSP
+import qualified Language.LSP.Types                     as LSP
 import           Language.LSP.VFS
-import           System.FilePath                      hiding (makeRelative)
+import           System.FilePath                        hiding (makeRelative)
 import           System.Time.Extra
 
 import           Data.IORef
@@ -143,13 +143,12 @@ import           OpenTelemetry.Eventlog
 import           PrelInfo
 import           UniqSupply
 
-import           Control.Exception.Extra              hiding (bracket_)
+import           Control.Exception.Extra                hiding (bracket_)
 import           Data.Default
 import           HieDb.Types
 import           Ide.Plugin.Config
-import qualified Ide.PluginUtils                      as HLS
-import           Ide.Types                            (PluginId)
-import           UnliftIO.Exception                   (bracket_)
+import qualified Ide.PluginUtils                        as HLS
+import           Ide.Types                              (PluginId)
 
 -- | We need to serialize writes to the database, so we send any function that
 -- needs to write to the database over the channel, where it will be picked up by
@@ -212,10 +211,6 @@ type WithProgressFunc = forall a.
     T.Text -> LSP.ProgressCancellable -> ((LSP.ProgressAmount -> IO ()) -> IO a) -> IO a
 type WithIndefiniteProgressFunc = forall a.
     T.Text -> LSP.ProgressCancellable -> IO a -> IO a
-
-data ProgressEvent
-    = KickStarted
-    | KickCompleted
 
 type GetStalePersistent = NormalizedFilePath -> IdeAction (Maybe (Dynamic,PositionDelta,TextDocumentVersion))
 
@@ -521,127 +516,6 @@ shakeOpen lspEnv defaultConfig logger debouncer
 
     return ideState
 
-data ProgressReporting  = ProgressReporting
-  { progressUpdate :: ProgressEvent -> IO ()
-  , inProgress     :: forall a. NormalizedFilePath -> Action a -> Action a
-  , progressStop   :: IO ()
-  }
-
-noProgressReporting :: IO ProgressReporting
-noProgressReporting = return $ ProgressReporting
-  { progressUpdate = const $ pure ()
-  , inProgress = const id
-  , progressStop   = pure ()
-  }
-
-delayedProgressReporting
-  :: Maybe (LSP.LanguageContextEnv c)
-  -> ProgressReportingStyle
-  -> IO ProgressReporting
-delayedProgressReporting lspEnv optProgressStyle = do
-    inProgressVar <- newVar HMap.empty
-    mostRecentProgressEvent <- newTVarIO KickCompleted
-    progressAsync <- async $
-            progressThread optProgressStyle mostRecentProgressEvent inProgressVar
-    let progressUpdate = atomically . writeTVar mostRecentProgressEvent
-        progressStop   = cancel progressAsync
-        inProgress :: NormalizedFilePath -> Action a -> Action a
-        inProgress = withProgressVar inProgressVar
-    return ProgressReporting{..}
-    where
-        -- The progress thread is a state machine with two states:
-        --   1. Idle
-        --   2. Reporting a kick event
-        -- And two transitions, modelled by 'ProgressEvent':
-        --   1. KickCompleted - transitions from Reporting into Idle
-        --   2. KickStarted - transitions from Idle into Reporting
-        progressThread style mostRecentProgressEvent inProgress = progressLoopIdle
-          where
-            progressLoopIdle = do
-                atomically $ do
-                    v <- readTVar mostRecentProgressEvent
-                    case v of
-                        KickCompleted -> STM.retry
-                        KickStarted   -> return ()
-                asyncReporter <- async $ mRunLspT lspEnv lspShakeProgress
-                progressLoopReporting asyncReporter
-            progressLoopReporting asyncReporter = do
-                atomically $ do
-                    v <- readTVar mostRecentProgressEvent
-                    case v of
-                        KickStarted   -> STM.retry
-                        KickCompleted -> return ()
-                cancel asyncReporter
-                progressLoopIdle
-
-            lspShakeProgress :: LSP.LspM config ()
-            lspShakeProgress = do
-                -- first sleep a bit, so we only show progress messages if it's going to take
-                -- a "noticable amount of time" (we often expect a thread kill to arrive before the sleep finishes)
-                liftIO $ sleep 0.1
-                u <- ProgressTextToken . T.pack . show . hashUnique <$> liftIO newUnique
-
-                void $ LSP.sendRequest LSP.SWindowWorkDoneProgressCreate
-                    LSP.WorkDoneProgressCreateParams { _token = u } $ const (pure ())
-
-                bracket_
-                  (start u)
-                  (stop u)
-                  (loop u 0)
-                where
-                    start id = LSP.sendNotification LSP.SProgress $
-                        LSP.ProgressParams
-                            { _token = id
-                            , _value = LSP.Begin $ WorkDoneProgressBeginParams
-                              { _title = "Processing"
-                              , _cancellable = Nothing
-                              , _message = Nothing
-                              , _percentage = Nothing
-                              }
-                            }
-                    stop id = LSP.sendNotification LSP.SProgress
-                        LSP.ProgressParams
-                            { _token = id
-                            , _value = LSP.End WorkDoneProgressEndParams
-                              { _message = Nothing
-                              }
-                            }
-                    sample = 0.1
-                    loop id prev = do
-                        liftIO $ sleep sample
-                        current <- liftIO $ readVar inProgress
-                        let done = length $ filter (== 0) $ HMap.elems current
-                        let todo = HMap.size current
-                        let next = 100 * fromIntegral done / fromIntegral todo
-                        when (next /= prev) $
-                          LSP.sendNotification LSP.SProgress $
-                          LSP.ProgressParams
-                              { _token = id
-                              , _value = LSP.Report $ case style of
-                                  Explicit -> LSP.WorkDoneProgressReportParams
-                                    { _cancellable = Nothing
-                                    , _message = Just $ T.pack $ show done <> "/" <> show todo
-                                    , _percentage = Nothing
-                                    }
-                                  Percentage -> LSP.WorkDoneProgressReportParams
-                                    { _cancellable = Nothing
-                                    , _message = Nothing
-                                    , _percentage = Just next
-                                    }
-                                  NoProgress -> LSP.WorkDoneProgressReportParams
-                                    { _cancellable = Nothing
-                                    , _message = Nothing
-                                    , _percentage = Nothing
-                                    }
-                              }
-                        loop id next
-
-        withProgressVar var file = actionBracket (f succ) (const $ f pred) . const
-            -- This functions are deliberately eta-expanded to avoid space leaks.
-            -- Do not remove the eta-expansion without profiling a session with at
-            -- least 1000 modifications.
-            where f shift = void $ modifyVar' var $ HMap.insertWith (\_ x -> shift x) file (shift 0)
-
 -- | Must be called in the 'Initialized' handler and only once
 shakeSessionInit :: IdeState -> IO ()
 shakeSessionInit IdeState{..} = do
@@ -801,18 +675,6 @@ instantiateDelayedAction (DelayedAction _ s p a) = do
           liftIO $ void $ try @SomeException $ signalBarrier b x
       d' = DelayedAction (Just u) s p a'
   return (b, d')
-
-mRunLspT :: Applicative m => Maybe (LSP.LanguageContextEnv c ) -> LSP.LspT c m () -> m ()
-mRunLspT (Just lspEnv) f = LSP.runLspT lspEnv f
-mRunLspT Nothing _       = pure ()
-
-mRunLspTCallback :: Monad m
-                 => Maybe (LSP.LanguageContextEnv c)
-                 -> (LSP.LspT c m a -> LSP.LspT c m a)
-                 -> m a
-                 -> m a
-mRunLspTCallback (Just lspEnv) f g = LSP.runLspT lspEnv $ f (lift g)
-mRunLspTCallback Nothing _ g       = g
 
 getDiagnostics :: IdeState -> IO [FileDiagnostic]
 getDiagnostics IdeState{shakeExtras = ShakeExtras{diagnostics}} = do
